feat: move util/has-prop to root shared folder
