+++ shared/util/has-prop.ts
