+++ src/app/modules/sgid/sgid.util.ts
@@ -1,6 +1,6 @@
 import { err, ok, Result } from 'neverthrow'
 
-import { hasProp } from '../../../shared/util/has-prop'
+import { hasProp } from '../../../../shared/util/has-prop'
 import {
   AuthType,
   BasicField,
